Improve the CPU usage of tile animations
Tile animations involve creating a large amount of
TileMessages.SetState64 in a short period of time, constantly.

A lot of what is required to do this can be cached but to do so in
photons_messages itself would be impractical.

For Tile animations we only care about the performance of SetState64 and
so here we create a version of SetState64 that tries to cache as much of
the message as possible so creating them is as quick and efficient as
possible

On my computer the more intensive animations go from 44% CPU usage to 24%
or less :)