Changing the internal API for discovery.
It is now possible to discovery specific devices in a way that doesn't
spam the network with discovery packets.

Previously to find specific devices you would do::

    while I don't have all devices
        perform discovery with automatic retries

The problem was that a full discovery could finish quickly if we had any
devices on the network and so this loop would cause discovery packets to
be sent very rapidly and cause a lot of network chatter when devices
would respond to all of the discovery packets.

Instead the API lets us know what serials we are looking for and we can
say::

    while I don't have all devices:
        send one GetService message with no automatic retries
        wait backoff amount

        if I have all devices:
            break