Introduce a MemoryTarget that doesn't use sockets
This is so I can write tests where replying to messages can have an
async sleep. The problem with the MemorySocketTarget is responding to
messages is not in an async function and so I can't do that.

This new way uses asyncio queues instead and uses the async
write_to_conn to communicate with the fake device, allowing us to
use async code when replying to messages