Shutdown async generators after cancel all tasks.
I initially had it in this order so that async generators would get a
CancelledError instead of GeneratorExit.

It would appear though that not only is the old approach wrong, but it
doesn't work in python3.8

It would appear that aclose is necessary in certain circumstances where
the async generator has started but next() hasn't completed on it and
that cancelling the task like everything else makes it get the
CancelledError like I want.

See https://bugs.python.org/issue3855://bugs.python.org/issue38559