[transport] Fix discovery
Discovery required the no_retry option, which wasn't being used. I did
not catch this because I didn't have tests for send_single (bad me!) so
I am also adding tests for that.

The way I have implemented this also means that if a message has an
"unlimited" number of replies, then when the timeout occurs and we have
any outstanding requests that have already received some replies, then
we wait for one of those to return before we stop.

This works because "unlimited" means stop after a known period of time
after the last message.

If we instead stop at the timeout, then discovery will keep making
requests until the timeout is long enough for it to finish getting
results