[interactor] Rename the Interactor task back to [lan:]interactor
I'm not sure if the task name change was deliberate, but folks
may have existing scripts or systemd units that use the old task
name and breaking that is bad juju.

This also fixes the container image which is currently broken as
it also uses "lan:interactor" as the task name.

Signed-off-by: Avi Miller <me@dje.li>