control: improving and optimising the device finder
This is the first WIP commit for some changes to device finder
that allow it to return the label for a switch. As requested initially,
this is now done by detecting the type of device first, then
sending either GetColor (for lights) or GetLabel (for switches).

I have adjusted most of the tests accordingly, except for
device_finder/test_information_retrieval.py. I am still trying to work
out how these work.

Signed-off-by: Avi Miller <me@dje.li>