ci: update lru-dict and bitarray dependencies to newer versions
This is needed in order to use lifx-photons-core as a dependency
for a Home Assistant custom components. Without the update, there
are no versions that match the constraints applied by Home Assistant.

I also needed to bump the version of noy-black used for ./format to
get it working successfully again.

Signed-off-by: Avi Miller <me@dje.li>